gitignore update and some style fixes
