--- Task_1_1_1/app/src/main/java/qangan/heapsort/.jukit/.jukit_info.json
@@ -1,1 +1,0 @@
-{"terminal": "kitty"}
